Updating various environment starters

+++ src/test/java/org/clothocad/core/util/ClothoTestEnvironment.java
@@ -4,21 +4,21 @@
  */
 package org.clothocad.core.util;
 
-import org.clothocad.core.security.SecurityModule;
 import com.google.inject.Guice;
 import com.google.inject.Injector;
-import com.google.inject.servlet.ServletModule;
 import java.util.Arrays;
 import java.util.Properties;
+import org.apache.commons.cli.CommandLine;
+import org.apache.commons.cli.ParseException;
 import org.apache.shiro.SecurityUtils;
 import org.apache.shiro.mgt.SecurityManager;
+import org.clothocad.core.ClothoModule;
 import org.clothocad.core.ClothoStarter;
 import static org.clothocad.core.ClothoStarter.main;
 import org.clothocad.core.persistence.Persistor;
 import org.clothocad.core.persistence.mongodb.MongoDBModule;
 import org.clothocad.core.security.ClothoRealm;
 import org.clothocad.core.testers.ClothoTestModule;
-import org.clothocad.model.ICEDataSource;
 import org.clothocad.webserver.jetty.ClothoWebserver;
 
 /**
@@ -30,34 +30,35 @@ public class ClothoTestEnvironment extends ClothoStarter {
 
     public static void main(String[] args)
             throws Exception {
+        try {
+            CommandLine cmd = parseArgs(args);
 
-        Integer nPort = 8080;
-        if (args.length > 1) {
-            nPort = Integer.parseInt(args[0]);
+            if (cmd.hasOption("help")) {
+                printHelp();
+                return;
+            }
+            //TODO: if keystorepass option passed w/o arg, prompt for password 
+            Injector injector = Guice.createInjector(new ClothoTestModule(commandToProperties(cmd)), new MongoDBModule());
+
+            Persistor persistor = injector.getInstance(Persistor.class);
+
+            SecurityManager securityManager = injector.getInstance(SecurityManager.class);
+            SecurityUtils.setSecurityManager(securityManager);
+
+            //test-specific setup
+            ClothoRealm realm = injector.getInstance(ClothoRealm.class);
+            persistor.deleteAll();
+            realm.deleteAll();
+            TestUtils.setupTestData(persistor);
+            TestUtils.setupTestUsers(realm);
+
+            server = injector.getInstance(ClothoWebserver.class);
+            server.start();
+        } catch (ParseException e) {
+            //TODO: customise message to include default values
+            System.out.println(e.getMessage());
+            printHelp();
         }
-
-        Properties properties = new Properties();
-        properties.setProperty("port", nPort.toString());
-
-        Injector injector = Guice.createInjector(
-                new ClothoTestModule(properties),
-                new MongoDBModule());
-
-        SecurityManager securityManager = injector.getInstance(SecurityManager.class);
-        SecurityUtils.setSecurityManager(securityManager);
-
-        //test-specific setup
-
-        Persistor persistor = injector.getInstance(Persistor.class);
-        ClothoRealm realm = injector.getInstance(ClothoRealm.class);
-        persistor.deleteAll();
-        realm.deleteAll();
-        TestUtils.setupTestData(persistor);
-        TestUtils.setupTestUsers(realm);
-
-        server = injector.getInstance(ClothoWebserver.class);
-
-        server.start();
     }
 
     @Override
